Split sources into separate directories.
