+++ src/DxLib/DxInternal.h
