+++ src/Makefile.am
@@ -2,48 +2,48 @@
 lib_LTLIBRARIES = libDxPortLib.la
 
 libDxPortLib_la_SOURCES =	\
-	DxDraw.c \
-	DxDXA.c \
-	DxFile.c \
-	DxFont.c \
-	DxGraph.c \
-	DxInternal.h \
-	DxLib.cpp \
-	DxLib_c.c \
-	Luna.cpp \
-	Luna3D.cpp \
-	LunaCamera.cpp \
-	LunaFile.cpp \
-	LunaFontSprite.cpp \
-	LunaInput.cpp \
-	LunaMath.cpp \
-	LunaRand.cpp \
-	LunaSound.cpp \
-	LunaSprite.cpp \
-	LunaTexture.cpp \
-	LunaVecMath.cpp \
-	PLAudio.c \
-	PLFile.c \
-	PLGLBuffers.c \
-	PLGLFixedFunction.c \
-	PLGLInternal.h \
-	PLGLRender.c \
-	PLGLTexture.c \
-	PLGLShaders.c \
-	PLHandle.c \
-	PLInput.c \
-	PLInternal.h \
-	PLMath.c \
-	PLRNG.c \
-	PLSDL2File.c \
-	PLSDL2GL.c \
-	PLSDL2Internal.h \
-	PLSDL2Main.c \
-	PLSDL2Window.c \
-	PLSDL2Memory.c \
-	PLSDL2SaveScreen.c \
-	PLSurface.c \
-	PLText.c \
-	PLText_CP932.c
+	DxLib/DxDraw.c \
+	DxLib/DxDXA.c \
+	DxLib/DxFile.c \
+	DxLib/DxFont.c \
+	DxLib/DxGraph.c \
+	DxLib/DxInternal.h \
+	DxLib/DxLib.cpp \
+	DxLib/DxLib_c.c \
+	Luna/Luna.cpp \
+	Luna/Luna3D.cpp \
+	Luna/LunaCamera.cpp \
+	Luna/LunaFile.cpp \
+	Luna/LunaFontSprite.cpp \
+	Luna/LunaInput.cpp \
+	Luna/LunaMath.cpp \
+	Luna/LunaRand.cpp \
+	Luna/LunaSound.cpp \
+	Luna/LunaSprite.cpp \
+	Luna/LunaTexture.cpp \
+	Luna/LunaVecMath.cpp \
+	PL/PLAudio.c \
+	PL/PLFile.c \
+	PL/PLHandle.c \
+	PL/PLInput.c \
+	PL/PLInternal.h \
+	PL/PLMath.c \
+	PL/PLRNG.c \
+	PL/GL/PLGLBuffers.c \
+	PL/GL/PLGLFixedFunction.c \
+	PL/GL/PLGLInternal.h \
+	PL/GL/PLGLRender.c \
+	PL/GL/PLGLTexture.c \
+	PL/GL/PLGLShaders.c \
+	PL/SDL2/PLSDL2File.c \
+	PL/SDL2/PLSDL2GL.c \
+	PL/SDL2/PLSDL2Internal.h \
+	PL/SDL2/PLSDL2Main.c \
+	PL/SDL2/PLSDL2Window.c \
+	PL/SDL2/PLSDL2Memory.c \
+	PL/SDL2/PLSDL2SaveScreen.c \
+	PL/PLSurface.c \
+	PL/PLText.c \
+	PL/PLText_CP932.c
 
 libDxPortLib_la_LDFLAGS = -version-info $(LT_CURRENT):$(LT_REVISION):$(LT_AGE)
