+++ src/PL/GL/PLGLTexture.c
@@ -23,7 +23,7 @@
 
 #ifdef DXPORTLIB_DRAW_OPENGL
 
-#include "PLInternal.h"
+#include "PL/PLInternal.h"
 
 #include "PLGLInternal.h"
