+++ src/PL/PLText.c
