+++ src/PL/PLText_CP932.c
